+++ files/gitlab-cookbooks/postgresql/recipes/directory_locations.rb
@@ -1,0 +1,3 @@
+node.default['gitlab']['postgresql']['unix_socket_directory'] ||= node['gitlab']['postgresql']['dir']
+node.default['gitlab']['postgresql']['data_dir'] ||= "#{node['gitlab']['postgresql']['dir']}/data"
+node.default['gitlab']['postgresql']['home'] ||= node['gitlab']['postgresql']['dir']